added multiple workers to project

+++ engine/color.cpp
@@ -1,6 +1,6 @@
 #include "color.h"
 
-Color::Color(int _r, int _g, int _b, int _a){
+Color::Color(int _r, int _g, int _b, float _a){
     this->r = _r;
     this->g = _g;
     this->b = _b;
