+++ engine/color.h
@@ -4,11 +4,11 @@
 class Color
 {
 public:
-    Color(int _r = 255, int _g = 255, int _b = 255, int _a = 255);
+    Color(int _r = 255, int _g = 255, int _b = 255, float _a = 255);
     int r;
     int g;
     int b;
-    int a;
+    float a;
 
 private:
